+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+### Added
+- Implement `Clone` and `Copy` for `BorrowedEntry`.
+- Implement `Clone` for `OwnedEntry`.
 
 ## [0.1.0] - 2024-04-13
 ### Added